Aligned model for upcoming lang/title select code.
