+++ Kavita.Models/DTOs/KavitaPlus/Metadata/SeriesRelationship.cs
@@ -4,13 +4,7 @@ using Kavita.Models.Entities.Enums.KavitaPlus;
 
 namespace Kavita.Models.DTOs.KavitaPlus.Metadata;
 
-public sealed record ALMediaTitle
-{
-    public string? EnglishTitle { get; set; }
-    public string RomajiTitle { get; set; }
-    public string NativeTitle { get; set; }
-    public string PreferredTitle { get; set; }
-}
+
 
 public sealed record SeriesRelationship
 {
